Move files to correct namespace
